Make the "unknown" type safe for functions
Previously, using the "unknown" type in function components was unsafe
and would lead to runtime crashes. By design, the runtime has no
information for how to deal with "unknown" data. This change makes it
safer by having it just return an internal representation of the data
when it encounters "unknown."

The representation is still unspecified, but it should be equal for
equal data. This is mainly useful for functions that just need to
compare arbitrary values.