Remove format specifier options
`%i` `%f` and `%b` do not take any flags now. The flags were not useful
in practice, they added a maintenance burden, and they were not flexible
enough to cover all use cases.

I may revisit this idea in the future. But, if they are added again,
they will need to be redesigned.