Reimplement the runtimes in tagless final style
This unifies the renderer and the JavaScript printer with a
tagless-final interface. Now both versions of the runtime use almost the
exact same semantics, and adding future runtimes will be easier. This
should not change the public interface except that some error messages
will be formatted differently to avoid the dependency on OCaml's Format.

This also bumps the required OCaml version to 5.0 so we can take
advantage of a few minor stdlib enhancements.