Add warnings & make API pure
This adds non-fatal warning functionality and converts a few
previously-fatal errors into warnings. It uses algebraic effects behind
the scenes, since effects are essentially resumable exceptions.

To avoid a complicated API that exposes both effects and exceptions, the
effect handlers are internal and handle the exceptions as well. As a
result, the public functions are now completely pure. Any function with
a possible error state returns a sequence of messages and an optional
return value.