Improve performance by reducing formatter calls
The OCaml format functions have significant performance costs at
runtime, so we only want to call them when necessary. However, each
formatted result may be used in multiple locations, and the
compile-to-JavaScript implementation shouldn't have to duplicate them.
This commit makes the formatters a first-class type in the semantics so
we can get the best of both worlds. Performance is measurably faster
compared to before.